finished base skeleton of graphql

+++ app/graphql/types/item_type.rb
@@ -2,7 +2,7 @@ module Types
   class ItemType < Types::BaseObject
     field :id, ID, null: false
     field :name, String, null: false
-    field :price, Decimal, null: false
+    field :price, Float, null: false
     field :image, String, null: false
     field :freeship, Boolean, null: false
     field :onsale, Boolean, null: false
